basic JSONHandle test are working
I realized that the AST json format is specific to `json-to-ast`.
Now I will experiment with the ts AST obtained from eslint.
I will start by creating a vscode command that open a terminal
and print the obtained json.

So before writing a navigator specific to the format of `json-to-ast`